chore: rename misspelled file entites.ts to entities.ts (#985)
